feat(FN-967): fix sending email and month calculation
