+++ portal-api/src/v1/controllers/utilisation-report-service/previous-reports.controller.js
@@ -1,12 +1,5 @@
 const api = require('../../api');
-
-const getMonthName = (monthNumber) => {
-  // date is set to 1st Jan to avoid bug when today's month has more days than target month
-  const date = new Date(2023, 1, 1);
-  // offset by 1 as January = 1 in Database
-  date.setMonth(monthNumber - 1);
-  return date.toLocaleString('default', { month: 'long' });
-};
+const { getMonthName } = require('../../../utils/getMonthName');
 
 const getYears = (reports) => {
   const years = reports.map((report) => report.year);
@@ -60,7 +53,6 @@ const getPreviousReportsByBankId = async (req, res) => {
 
 module.exports = {
   getPreviousReportsByBankId,
-  getMonthName,
   getYears,
   getReportsGroupedByYear,
   populateOmittedYears,
