+++ portal-api/src/v1/controllers/utilisation-report-service/previous-reports.controller.test.js
@@ -82,61 +82,6 @@ describe('controllers/utilisation-report-service/previous-reports', () => {
     reports: [],
   }];
 
-  it.each([
-    {
-      monthNumber: 1,
-      monthName: 'January',
-    },
-    {
-      monthNumber: 2,
-      monthName: 'February',
-    },
-    {
-      monthNumber: 3,
-      monthName: 'March',
-    },
-    {
-      monthNumber: 4,
-      monthName: 'April',
-    },
-    {
-      monthNumber: 5,
-      monthName: 'May',
-    },
-    {
-      monthNumber: 6,
-      monthName: 'June',
-    },
-    {
-      monthNumber: 7,
-      monthName: 'July',
-    },
-    {
-      monthNumber: 8,
-      monthName: 'August',
-    },
-    {
-      monthNumber: 9,
-      monthName: 'September',
-    },
-    {
-      monthNumber: 10,
-      monthName: 'October',
-    },
-    {
-      monthNumber: 11,
-      monthName: 'November',
-    },
-    {
-      monthNumber: 12,
-      monthName: 'December',
-    },
-  ])('should return month name from month number', (value) => {
-    const monthName = getMonthName(value.monthNumber);
-
-    expect(monthName).toEqual(value.monthName);
-  });
-
   it('should return unique array of years', () => {
     const uniqueYears = getYears(reports);
